+++ atomic/atom/smfetch/src/browser/smfetch/atom.js
@@ -166,6 +166,7 @@ export default await (() => {
           error: fault,
           ajax = true,
           option = {},
+          download = false,
         } = param;
 
         let {
@@ -196,6 +197,7 @@ export default await (() => {
                 if (achieve) {
                   let result = {};
                   if (response.redirected) result.redirected = response.url;
+                  else if (download) result.data = await response.blob();
                   else result.data = await response.json();
                   success({
                     status: response.status,
@@ -229,6 +231,8 @@ export default await (() => {
           if (response.ok) {
             if (response.redirected) {
               result.redirected = response.url;
+            } else if (download) {
+              result.data = await response.blob();
             } else {
               let resp = await response.json();
               if (ajax) result.data = await resp;
@@ -275,7 +279,6 @@ export default await (() => {
         if (param?.success !== undefined) data.success = param.success;
         if (param?.error !== undefined) data.error = param.error;
         if (param?.async !== undefined) async = param.async;
-        if (param?.reroute !== undefined) data.reroute = param.reroute;
 
         if (async) {
           window.fetchapi.request(data);